use sparkfun fork of roadrunner

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc7083/autonomous/drive/ThreeDeadWheelLocalizer.java
@@ -24,7 +24,7 @@ public final class ThreeDeadWheelLocalizer implements Localizer {
     public static Params PARAMS = new Params();
     public final Encoder par0, par1, perp;
     public final double inPerTick;
-    private int lastPar0Pos, lastPar1Pos, lastPerpPos;
+    private double lastPar0Pos, lastPar1Pos, lastPerpPos;
     private boolean initialized;
 
     public ThreeDeadWheelLocalizer(HardwareMap hardwareMap, double inPerTick) {
@@ -58,9 +58,9 @@ public final class ThreeDeadWheelLocalizer implements Localizer {
             );
         }
 
-        int par0PosDelta = par0PosVel.position - lastPar0Pos;
-        int par1PosDelta = par1PosVel.position - lastPar1Pos;
-        int perpPosDelta = perpPosVel.position - lastPerpPos;
+        double par0PosDelta = par0PosVel.position - lastPar0Pos;
+        double par1PosDelta = par1PosVel.position - lastPar1Pos;
+        double perpPosDelta = perpPosVel.position - lastPerpPos;
 
         Twist2dDual<Time> twist = new Twist2dDual<>(
                 new Vector2dDual<>(
